Refactor type hints and add GuzzleException throws
Updated type hints in several methods to reflect correct data types, such as changing `float` to `int` where necessary. Added GuzzleException throws annotations in APIFuturePosting. This enhances code robustness and improves API error handling.

+++ src/Common/Product.php
@@ -17,7 +17,7 @@ class Product
     private ?string $reference;
     private ?string $productCode;
     private ?string $productName;
-    private ?float $productQty;
+    private ?int $productQty;
     private ?float $productValue;
     private ?int $automaticType;
     private ?int $accountType;
